+++ v2/Include/System/Core/eosAllocator.h
@@ -6,17 +6,13 @@
 //
 #include "eos.h"
 
-// Standard includes
-//
-#include "stddef.h"
-
 
 namespace eos {
 
     class IAllocator {
         public:
     		virtual ~IAllocator() {}
-            virtual void *allocate(size_t size) = 0;
+            virtual void *allocate(unsigned size) = 0;
             virtual void deallocate(void *p) = 0;
     };
